auth and methods for scores and ranking interfaces

+++ src/main/java/com/pfa/pfaproject/dto/Dimension/GetYearDimensionsDTO.java
@@ -4,6 +4,6 @@ public record GetYearDimensionsDTO(
         Long id,
         String name,
         String description,
-        double weight
+        Integer weight
 ) {
 }
